requirements.txt
Flask>=2.3
PyYAML>=6.0
gunicorn>=21.0
supabase>=2.20
websockets>=12.0